Fixed packaging and changed get_nickname to base off of email instead of fullname
