+++ app/__init__.py
@@ -1,2 +1,3 @@
 from . import settings
-from .utility_functions import *
+from .utility import *
+from .gui import App
