+++ html_templates/offices/acceptance/marketing_acceptance.html
@@ -114,7 +114,7 @@
                                         Your insight and passion have caught our eye, and we're beyond excited to have
                                         you on board. As a valued member of our
                                         <span class="yellow">Marketing Office</span>, you'll be under the mentorship of
-                                        our esteemed <span class="purple">Chief Marketing Officer (CMO)</span>, Ms.
+                                        our chootie <span class="purple">Chief Marketing Officer (CMO)</span>, Ms.
                                         Clareene Roterfil D. Borromeo, and
                                         <span class="purple">Vice-Chief Marketing Officer (VCMO)</span>, Mr. Sean Jakob
                                         C. Mejino!
